creation of quote_card to avoid repetition for many pages - fix the rating on the quote - change the season/id page

+++ src/Entity/Quote.php
@@ -25,7 +25,7 @@ class Quote
     private $text;
 
     /**
-     * @ORM\Column(type="smallint", nullable=true)
+     * @ORM\Column(type="float", nullable=true)
      */
     private $rating;
 
@@ -82,12 +82,12 @@ class Quote
         return $this;
     }
 
-    public function getRating(): ?int
+    public function getRating(): ?float
     {
         return $this->rating;
     }
 
-    public function setRating(?int $rating): self
+    public function setRating(?float $rating): self
     {
         $this->rating = $rating;
